+ ♻️ significantly improved badge list (#141)
Many things have been either added, removed, or adjusted

✨ Added:
* Bitcoin Cash
* Bitcoin SV
* Stellar
* Neovim
* JSON
* Proton Mail

🗑️ Removed:
* Web Technologies (Category)
* Version Control (Category)

To be in line with mdbadges-cli, I removeds these categories as, in web technolgies' case, it was basically a bunch of the same badges. For Version Control, it was too much to add to mdbadges-cli, and as such, has been removed.

♻️ Changed:
* Dev.to
* Safari
* Firebase (Cloud Systems)
* Stellar
* MySQL
* PayPal
* Ember -> Ember.js
* TailwindCSS -> Tailwind CSS
* Tumblr (Social Media)